housekeeping and cleanup
