+++ map/testmap.cub
@@ -1,0 +1,11 @@
+	111111111111111
+	100000000000101
+	100001000000101
+	111100000010101
+	100000000010101
+	100000001111101
+	100000000000001
+	100000000000001
+	111111000111101
+	100000000000001
+	111111111111111
