+++ src/draw_line.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   draw_line.c                                        :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: jdagoy <jdagoy@student.42.fr>              +#+  +:+       +#+        */
+/*   By: jdagoy <jdagoy@student.s19.be>             +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/12/18 15:01:56 by jdagoy            #+#    #+#             */
-/*   Updated: 2023/12/29 15:16:02 by jdagoy           ###   ########.fr       */
+/*   Updated: 2024/01/03 14:28:53 by jdagoy           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -63,45 +63,4 @@ void	draw_line_dda(t_image_data *img, t_point start, t_point end)
 		pixel.axis[X] += indices[X];
 		pixel.axis[Y] += indices[Y];
 	}
-}
-
-void	draw_grids(t_display *mlx)
-{
-	int		axis[2];
-	t_point	end;
-
-	axis[Y] = 0;
-	while (axis[Y] < WIDTH / 2)
-	{
-		end = create_point(axis[Y], HEIGHT);
-		end.color = GRAY;
-		draw_line_dda(&mlx->img, create_point(axis[Y], 0), end);
-		axis[Y] += TILE_SIZE;
-	}
-	axis[X] = 0;
-	while (axis[X] < HEIGHT)
-	{
-		end = create_point(WIDTH / 2, axis[X]);
-		end.color = GRAY;
-		draw_line_dda(&mlx->img, create_point(0, axis[X]), end);
-		axis[X] += TILE_SIZE;
-	}
-}
-
-t_point	create_point(float x, float y)
-{
-	t_point	point;
-
-	point.axis[X] = x;
-	point.axis[Y] = y;
-	return (point);
-}
-
-t_vector	create_vector(float x, float y)
-{
-	t_vector	vector;
-
-	vector.dir[X] = x;
-	vector.dir[Y] = y;
-	return (vector);
 }
