+++ src/init_player.c
@@ -1,0 +1,27 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   init_player.c                                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jdagoy <jdagoy@student.s19.be>             +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2024/01/03 14:24:11 by jdagoy            #+#    #+#             */
+/*   Updated: 2024/01/03 14:36:06 by jdagoy           ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
+#include "cub3d.h"
+
+
+void	update_player_front(t_player *player)
+{
+	player->front.dir[X] = cos(player->angle) * 5;
+	player->front.dir[Y] = (sin(player->angle) * 5) * -1;
+}
+
+void	init_player(t_player *player)
+{
+	player->position = create_point((WIDTH / 2) / 2, HEIGHT / 2);
+	player->angle = M_PI / 2;
+	update_player_front(player);
+}
